MB-10057: implemented vbuuid creation
The trick we use is to update vbucket database during mccouch
notification. At this time we know that ep-engine does not touch
.couch file.

Notable implementation detail is that we have to manually spawn
duplicate couchdb process for actual update. That's because normal
couchdb instance we have for normal vbuckets refuse any updates (which
is a feature).

Change-Id: I1e245ef0462cfc45d282fe66c78f9a80f97e582c
Reviewed-on: http://review.couchbase.org/34038
Reviewed-by: Aliaksey Artamonau <aliaksiej.artamonau@gmail.com>
Tested-by: Aliaksey Kandratsenka <alkondratenko@gmail.com>