MB-9979 An option to require rebalance to perform delta recovery.
By default delta recovery is performed for requested nodes if it's
possible. If it's not, normal rebalance is performed. With this option
rebalance will fail to start if delta recovery is not possible. So
that user could take any actions (like adding or removing nodes)
for delta recovery to be possible.

Change-Id: I7b5e0d9914dccb584c3ef8997267b0189b4be1b7
Reviewed-on: http://review.couchbase.org/34223
Tested-by: Aliaksey Artamonau <aliaksiej.artamonau@gmail.com>
Reviewed-by: Aliaksey Kandratsenka <alkondratenko@gmail.com>