Improve rebalance error reporting.
As far as I can tell, unknownNodes and {mismatch: 1} errors are there
for the same reason and depending on circumstances one or the other
can be returned. For example, if user passed an invalid node name that
happened to coincide with some existing atom, then "mismatch" error
would be returned. Otherwise, "unknownNodes" would be returned.

Our management UI handles only "mismatch" error and actually only this
error is documented in our API reference, so I removed the other
one. Even if any client was actually aware of this error, it must have
also been aware of the other one, so nothing should break.

Change-Id: I418d648806cb7f53732ec44f5384cf9a646e4fe2
Reviewed-on: http://review.couchbase.org/34271
Reviewed-by: Aliaksey Kandratsenka <alkondratenko@gmail.com>
Tested-by: Aliaksey Artamonau <aliaksiej.artamonau@gmail.com>