implemented support of delta recovery of specific buckets
requireDeltaRecovery query parameter is now not used and is
ignored. We now always fail if delta recovery for all specified (via
setRecoveryType as usual) nodes is impossible. I.e. as if
requireDeltaRecovery=true is passed.

And we now have deltaRecoveryBuckets POST parameter which can be used
to pass list of buckets that need to be recovered. It contains comma
separated list of bucket names. When requireDeltaRecoveryBuckets is
passed, then only specified buckets will be recovered on all nodes
marked for delta recovery. And if at least one node for at least one
bucket in this list cannot be delta-recovered, then entire operation
will be refused.

When requireDeltaRecoveryBuckets is not passed it is assumed that all
buckets need to be recovered.

Change-Id: Iaac7b8e7000a8eeb85cc5a9115410e5802aa1161
Reviewed-on: http://review.couchbase.org/35222
Tested-by: Aliaksey Kandratsenka <alkondratenko@gmail.com>
Reviewed-by: Aliaksey Artamonau <aliaksiej.artamonau@gmail.com>