always immediately consume initial chunk of data from upr
It's possible that "leftover" from receiving initial response has upr
data and end stream message. In which case we'd enter socket loop
waiting for more data in order to process that leftover. But because
in this case leftover is all we would ever have we'd stuck.

The fix is to handle leftover as if we've just received it.

Change-Id: I57c0b7661595ecc23a702dbe156ebb1b7c0fb41b
Reviewed-on: http://review.couchbase.org/35429
Tested-by: Aliaksey Kandratsenka <alkondratenko@gmail.com>
Reviewed-by: Artem Stemkovski <artem@couchbase.com>