don't crash connection pool when handling done from unknown pid
There's a tiny chance for connection pool to receive EXIT message and
die when xdcr is having frequent crashes (i.e. during
rebalance). That's relatively normal.

What is not normal is that newly started connection pool instance
would crash again when socket is returned back to it from unknown
process.

In order to fix it we now simply drop such requests (and sockets) on
the floor.

Change-Id: I5cfe87c88f8f7b497eb9c89e9846b0fb871c3b94
Reviewed-on: http://review.couchbase.org/35530
Tested-by: Aliaksey Kandratsenka <alkondratenko@gmail.com>
Reviewed-by: Artem Stemkovski <artem@couchbase.com>