MB-10858 Start ns_server_sup via intermediate process.
The deletions of per node keys when node gets ejected can be
replicated to this node after comes back online. This may prevent some
of the ns_server_sup kids from starting normally. And because of the
way supervisor handles start up errors, it causes all the children of
both ns_server_sup and ns_server_cluster_sup to be shut down and then
restarted. This prevents ns_cluster from completing the leave because
it also gets stopped as a kid of ns_cluster_server_sup.

So the fix is to start ns_server_sup via a wrapper process which
pretends that ns_server_sup starts immediately. So supervisor doesn't
shutdown everything and ns_cluster can complete its job.

Change-Id: Ieb5ff0fc6318f8660b781cbeebad6a26563b7226
Reviewed-on: http://review.couchbase.org/36020
Tested-by: Aliaksey Artamonau <aliaksiej.artamonau@gmail.com>
Reviewed-by: Artem Stemkovski <artem@couchbase.com>