MB-9915: redirect capi requests if node doesn't host given bucket
There are cases where bucket does exist in cluster and node is aware
of it. So it can auth request. But node is not in list of bucket nodes
(yet).

In this case attempt to perform any capi requests for given bucket on
this node is futile. Even master vbucket does not exist on this node
(yet).

So just like for views with node vbuckets on given node we simply try
to redirect to a node that (probably still) does host bucket.

Change-Id: Ia0fcf1b02dd018fecb26935f451495371043a62d
Reviewed-on: http://review.couchbase.org/36805
Tested-by: Aliaksey Kandratsenka <alkondratenko@gmail.com>
Reviewed-by: Aliaksey Artamonau <aliaksiej.artamonau@gmail.com>