MB-11124 Bump maximum number of ets tables.
Default value for this is 2053. Since every xdc_vbucket_rep creates an
ets table, it's barely enough to have two outgoing xdcr replications
in the worst case.

Erlang preallocates an array of slots for all the ets tables. Size of
each slot is sizeof(void *) on the system. So increase in memory
consumption isn't that high.

Change-Id: I8e7d568f780e37f4e2d74d75cbc295ae2bfaa66e
Reviewed-on: http://review.couchbase.org/37401
Tested-by: Aliaksey Artamonau <aliaksiej.artamonau@gmail.com>
Reviewed-by: Aliaksey Kandratsenka <alkondratenko@gmail.com>