MB-11228: handle end_stream right after stream_request
In some cases (like upr_notifier failing, for example if bucket
instance does not exist yet) we might be woken up to replicate even if
xdcr is fully up to date.

In that case we'll ask ep-engine to get us upr stream with start seqno
and end seqno being same numbers. And in that case ep-engine will not
send us upr_snapshot_marker but it'll give us stream_end immediately
(after checking for rollback as usual).

We now handle this case by faking and handling trivial snapshot
marker.

Change-Id: Ie13d1e76c9654ec3fb7972f9a91eb6b9996c8f75
Reviewed-on: http://review.couchbase.org/37695
Reviewed-by: Aliaksey Artamonau <aliaksiej.artamonau@gmail.com>
Tested-by: Aliaksey Kandratsenka <alkondratenko@gmail.com>