Handle more http versions in access.log formatter.
If version is omitted in the request it is assumed to be 0.9 which
would crash the formatter. So I added a case for this. I also added a
default case, in case someone by mistake or on purpose sends other
version to us. We used to be permissive about such requests, this
allows us to continue handling them.

Change-Id: I4c2fafa40004d9493fb9b2931c2acc1fa6333f3c
Reviewed-on: http://review.couchbase.org/37972
Tested-by: Aliaksey Artamonau <aliaksiej.artamonau@gmail.com>
Reviewed-by: Aliaksey Kandratsenka <alkondratenko@gmail.com>