MB-11370: implement proper erlang arg escaping for cluster_run
We also revert "MB-11370: Revert "pass nodefile arg and absolutize
datapath..."". This reverts commit
669831ca7ec30febcbd0203d72773187d9cb3247.

A number of erlang args are parsed as erlang terms. So when we pass
some path we have to pass it as readable representation of erlang's
string term. New code should do the trick (assuming nobody will bother
us with paths that contain newlines etc).

Change-Id: Ic9466645a609df2b1ec0c518616f80eb58630ea3
Reviewed-on: http://review.couchbase.org/38094
Reviewed-by: Aliaksey Artamonau <aliaksiej.artamonau@gmail.com>
Tested-by: Aliaksey Artamonau <aliaksiej.artamonau@gmail.com>