MB-11295 Refetch remote cluster when reference changes.
Specifically when node from reference is not in the cached list of
nodes. Note that we do this only when remote bucket is fetched and
there's no cached version of it. But that should be fine, because if
cached version is inconsistent and causes replicator to crash, it will
invalidate the cache.

Change-Id: I551aba0ec22e270f3fdd2145be5663231b99c7fe
Reviewed-on: http://review.couchbase.org/38277
Tested-by: Aliaksey Artamonau <aliaksiej.artamonau@gmail.com>
Reviewed-by: Aliaksey Kandratsenka <alkondratenko@gmail.com>