initial json xdcr tracing implementation
We're replacing human-readable xdcr_trace log with machine readable
trace. The intention is to keep tracing disabled. With ability to
enable it at runtime.

And intention is that it must be possible to recover exactly what was
done by xdcr. Which should help figuring out timings but most
importantly to help debugging conflict resolution bugs in xdcr code or
in ep-engine.

Change-Id: Ib5895c7fbc7b2de9b9a75236e88feab1ade1b731
Reviewed-on: http://review.couchbase.org/38450
Tested-by: Aliaksey Kandratsenka <alkondratenko@gmail.com>
Reviewed-by: Artem Stemkovski <artem@couchbase.com>