MB-10954 Use needs_rebalance when deciding on failover warnings.
So that we (almost) always show a warning when rebalance button is
enabled. The difference is in the set of nodes we pass to the
function. For the sake of failover safety warnings we pass only active
nodes. This helps us avoid a duplication of information that's shown to
the user. Specifically, otherwise we'd show a warning when the user just
adds a node to a cluster without hitting rebalance.

Change-Id: Ib5627b51919db2e7c1077f18d74e1c78e755f76b
Reviewed-on: http://review.couchbase.org/38280
Tested-by: Aliaksey Artamonau <aliaksiej.artamonau@gmail.com>
Reviewed-by: Aliaksey Kandratsenka <alkondratenko@gmail.com>