MB-11182: don't fail entire cbcollect_info if cbstats is missing
Popen may raise exception of spawning child fails. In most cases we
spawn shell which usually works. So we aren't seeing it in practice.

However few programs, particularly cbstats, are spawned not via
shell. In that case lack of program would indeed cause Popen to raise
exception which in previous code caused entire cbcollect_info to fail.

We handle it now in a way similar to how shell handles it.

Change-Id: I051c830ade62b9cc5f8706feeae0d16a3875ebdc
Reviewed-on: http://review.couchbase.org/38706
Tested-by: Aliaksey Kandratsenka <alkondratenko@gmail.com>
Reviewed-by: Aliaksey Artamonau <aliaksiej.artamonau@gmail.com>