disable xdcr datatype support completely
It's almost certain that 3.0 will not support datatype extension. And
also there's still chance that 3.1 datatype extension will be
different than 3.0. Making it unsafe to replicate 3.0's datatypes to
3.1.

In currently 3.0, even with datatype disabled, memcached still does
json detection. So we can see datatype 1 (json) in upr. This change
makes sure that we drop datatype to 0 for xdcr replication. Regardless
of advertised datatype support on destination.

Change-Id: Ia3064c233d08428c1adf14dfd696566d5bbfd7ee
Reviewed-on: http://review.couchbase.org/38701
Reviewed-by: Aliaksey Artamonau <aliaksiej.artamonau@gmail.com>
Tested-by: Aliaksey Kandratsenka <alkondratenko@gmail.com>