set up memcached to maintain less top keys
Because with our previous setting we had to consume 800 top keys every
time time we asked them. And that's despite us needing just to 10 per
node.

New setting is 40 which is still more than 10, but we need to give
internal LRU that's used for top keys calculating at least a tiny bit
of space above that count that we actually going to use.

Change-Id: Ib5942dd05cae43a78649145ab9c39f9475575145
Reviewed-on: http://review.couchbase.org/38891
Tested-by: Aliaksey Kandratsenka <alkondratenko@gmail.com>
Reviewed-by: Aliaksey Artamonau <aliaksiej.artamonau@gmail.com>