prevent attempts to open vbuckets in couchdb
As part of investigating MB-11751 I saw a bunch of vbuckets being open
in couchdb. For no apparent reason (specifically, no signs of geocouch
ever used). And I was unable to reproduce this condition.

So in order to catch and remaining offenders I want to merge this
change which will refuse any attempts to open vbuckets.

Change-Id: I08547d29d94d425263370cb3d84dacd8c1270c49
Reviewed-on: http://review.couchbase.org/39474
Reviewed-by: Sarath Lakshman <sarath@couchbase.com>
Tested-by: Aliaksey Kandratsenka <alkondratenko@gmail.com>