MB-8746: Use the set view based spatial index
With this commit we switch to using the set view based spatial index.

The configuration file for it now lives in the GeoCouch project instead
of in ns_server/etc/geocouch.ini. This is also where the hooks to
`capi_spatial` are defined (in case you wonder why there's no direct
trace of it when grepping for it in ns_server.

Change-Id: I672d9e24f84f4c3ec6891e3e1789d740415fd63d
Reviewed-on: http://review.couchbase.org/40117
Reviewed-by: Aliaksey Kandratsenka <alkondratenko@gmail.com>
Reviewed-by: Volker Mische <volker.mische@gmail.com>
Tested-by: Volker Mische <volker.mische@gmail.com>