Add license to erlwsh
In order to be able to package Couchbase properly for Debian, all
components must have proper license information. The upstream erlwsh
added the Apache License.

In order to make things as easy as possible in regards to licensing,
it makes sense to update to the latest upstream version. This commit
does exactly that. The only difference between the upstream version
and our's after this commit is merged will be a single line change in
`src/erlwsh_web` where we use a different path.

Change-Id: Ic70ca02c06a0710cb5e07ed414d0169e0fd82233
Reviewed-on: http://review.couchbase.org/40746
Reviewed-by: Aliaksey Kandratsenka <alkondratenko@gmail.com>
Tested-by: Aliaksey Kandratsenka <alkondratenko@gmail.com>