bp: MB-12147 Check all required stats for disk/data usage.
We have couch_disk_data_size (as well as some other stats that don't
make sense) even if there're only memmcached buckets. That's because in
certain places we don't make clear separation between memcached and
membase buckets in stats_collector code. I decided that it's easier to
just check all the required stats and fix the referred bug in this way.

Change-Id: Ia1620031aa87245178e205a30f7e97ded2ac7451
Orig-Reviewed-on: http://review.couchbase.org/41379
Reviewed-on: http://review.couchbase.org/42988
Reviewed-by: Aliaksey Kandratsenka <alkondratenko@gmail.com>
Tested-by: Aliaksey Kandratsenka <alkondratenko@gmail.com>