MB-12147 Check all required stats for disk/data usage.
We have couch_disk_data_size (as well as some other stats that don't
make sense) even if there're only memmcached buckets. That's because in
certain places we don't make clear separation between memcached and
membase buckets in stats_collector code. I decided that it's easier to
just check all the required stats and fix the referred bug in this way.

Change-Id: I3f17971f2c532c31e63237eb2f5c6956df95e98d
Reviewed-on: http://review.couchbase.org/41379
Reviewed-by: Aliaksey Kandratsenka <alkondratenko@gmail.com>
Tested-by: Aliaksey Kandratsenka <alkondratenko@gmail.com>