MB-12158: don't trap exits in dcp_proxy processes
Because we don't really need to.

And most importantly because erlang has a bug where send might get
stuck forever if socket is closed on other end.

So if we just die when dcp_replicator dies we'll avoid this
problem. Or at very least we'll close both sockets without any delays.

Change-Id: I4ea2225ea78a5543053396574d33f8459d22c302
Reviewed-on: http://review.couchbase.org/41410
Reviewed-by: Aliaksey Kandratsenka <alkondratenko@gmail.com>
Tested-by: Aliaksey Kandratsenka <alkondratenko@gmail.com>