bp:MB-12197: made terminate_and_wait work even for normal reason
When exit is send with reason normal, it gets ignored if child is not
trapping exits.

Recent change of dcp_replicator childs to not trap exits caused
condition where terminate with reason normal got stuck because of
normal exit being ignored.

This commit maps normal to shutdown and fixes MB.

Change-Id: Id02fdbbeebd8bc225ada8a849b7940b4301a7bbd
Orig-Reviewed-on: http://review.couchbase.org/41474
Reviewed-on: http://review.couchbase.org/42399
Reviewed-by: Aliaksey Kandratsenka <alkondratenko@gmail.com>
Tested-by: Aliaksey Kandratsenka <alkondratenko@gmail.com>