MB-12197: made terminate_and_wait work even for normal reason
When exit is send with reason normal, it gets ignored if child is not
trapping exits.

Recent change of dcp_replicator childs to not trap exits caused
condition where terminate with reason normal got stuck because of
normal exit being ignored.

This commit maps normal to shutdown and fixes MB.

Change-Id: I913814746f24446532fc0462b208f527993f9900
Reviewed-on: http://review.couchbase.org/41474
Tested-by: Aliaksey Kandratsenka <alkondratenko@gmail.com>
Reviewed-by: Aliaksey Artamonau <aliaksiej.artamonau@gmail.com>