MB-12382: unbreak config upgrade from 2.x to 3.0.1
When upgrading config from 2.3 to 3.0 the code that is supposed to
rewrite verbosity is not working anymore. Due to recent removal of
verbosity attribute from memcached config.

I decided not to change old config upgrade but to patch things up in
code that upgrades config to 3.0.1. New code simply removes verbosity
attribute from memcached config.

Change-Id: I1e9fc31f880257d7da7d2b027550b56540d04ff4
Reviewed-on: http://review.couchbase.org/42299
Tested-by: Aliaksey Kandratsenka <alkondratenko@gmail.com>
Reviewed-by: Aliaksey Artamonau <aliaksiej.artamonau@gmail.com>