MB-12866: Include breakpad minidumps in cbcollect_info
Expand dump_guts to return the directory that Breakpad saves minidumps
to, then use this in cbcollect_info to add any .dmp files in that
directory to cbcollect_info.

Note that the collection explicitly doesn't use the existing Task
mechanism, mainly as we just need the raw files so there's no point in
copying them; plus it is useful to preserve the creation date of the
dump files.

Change-Id: I311f99d14d513f098d5f23b24bfcadd70264e594
Reviewed-on: http://review.couchbase.org/45395
Reviewed-by: Aliaksey Artamonau <aliaksiej.artamonau@gmail.com>
Tested-by: Aliaksey Kandratsenka <alkondratenko@gmail.com>