Throw errors in mc_binary:recv_with_data.
In most cases we just assert that its result is of the form {ok, _,
_}. But it complicates matters if we actually want to either handle an
error or just pass it in a comprehensible form to the caller. This is
exactly the case with get_keys where we don't want to crash in case of
error, but merely pass it to the caller. We already have a try...catch
around do_get_keys, so changing to throw almost automagically does the
job.

Change-Id: If1b82dfd0a83e1c89f9cb80f3ecf368d499e1f20
Reviewed-on: http://review.couchbase.org/45651
Tested-by: Aliaksey Artamonau <aliaksiej.artamonau@gmail.com>
Reviewed-by: Aliaksey Kandratsenka <alkondratenko@gmail.com>