Restart only ns_server_nodes_sup when changing paths.
Actually, it should be enough to restart the couchdb. But to play it
safe, we continue restarting most of our processes.

This change is meant purely to simplify former somewhat confusing code
to restart ns_server app (with the exception of dist_manager). Nothing
in ns_server_cluster_sup (with the exception of ns_server_nodes_sup)
depends on these paths, so it's safe to ignore rest of the processes
there.

Change-Id: Icf854ffcaf9fc1b66d9e8381a7c5ac77e56ce514
Reviewed-on: http://review.couchbase.org/47479
Reviewed-by: Artem Stemkovski <artem@couchbase.com>
Tested-by: Artem Stemkovski <artem@couchbase.com>