A wrapper process to safely restart our subcomponents.
It's basically meant to replace the sequence of
supervisor:terminate_child and supervisor:restart_child. The problem
with the latter is that if your process fails in between the two
calls, nobody will try restarting the subcomponent.

The wrapper sits in between a supervisor and the actual child
process. It propagates exit signals in both directions and also
handles restart calls. If something goes wrong in the middle of
restart, then the wrapper itself will crash and the parent supervisor
will try to restart it. Which in turn will restart the child process.

Change-Id: I6318f58329dd6d47c43005a36154d4178f3d2554
Reviewed-on: http://review.couchbase.org/47480
Reviewed-by: Artem Stemkovski <artem@couchbase.com>
Tested-by: Artem Stemkovski <artem@couchbase.com>