MB-13567 Make design documents available on non-kv nodes.
Note, that as of now views still cannot be queried. But the list of
design documents is available on non-kv nodes as soon as they are
rebalanced in. Also new design documents can be created on non-kv
nodes too.

I had to split single_bucket_sup in two supervisors:
single_bucket_kv_sup and docs_sup. And the docs_sup itself into
docs_sup and docs_kv_sup. The _kv_ infix indicates that such
supervisor is running only on kv nodes. docs_sup is running on non-kv
nodes after they are rebalanced in.

One more missing piece (as of now) is master vbucket compaction.

Change-Id: I0763569dd26a05d0d795e91af478339f17085add
Reviewed-on: http://review.couchbase.org/48045
Tested-by: Aliaksey Artamonau <aliaksiej.artamonau@gmail.com>
Reviewed-by: Artem Stemkovski <artem@couchbase.com>