fix race in cluster_compat_mode:is_goxdcr_enabled
It's not possible to atomically update several keys in ets table.
So even though stop_xdcr and cluster_compat_version are updated
in a single transaction, they it's possible to see the state where
one is updated but not the other one.

This fix makes sure that goxdcr won't start if such race condition occur.

Change-Id: I0a563378c387d93f6af9a8984a9d4f893e370135
Reviewed-on: http://review.couchbase.org/48262
Reviewed-by: Aliaksey Artamonau <aliaksiej.artamonau@gmail.com>
Tested-by: Aliaksey Artamonau <aliaksiej.artamonau@gmail.com>