quotaTotal and quotaUsed consider all the nodes
Not just actually alive ones. Old behavior is just an artifact of the
way this code was written initially. Then stats (not ns_doctor) were
used to get all the needed information. Which meant that only live
nodes had to be considered. Then this code was changed many times and
this behavior was preserved only for the aforementioned two fields.

Change-Id: Iad9cebf17d6acb5ba6899ec070e8832d1d5ebc4a
Reviewed-on: http://review.couchbase.org/49138
Tested-by: Aliaksey Artamonau <aliaksiej.artamonau@gmail.com>
Reviewed-by: Artem Stemkovski <artem@couchbase.com>