MB-13615 Always attempt to restart ns_server after join.
The logic with markers relies on the fact that if something goes wrong,
then ns_cluster just crashes. But perform_actual_join can actually fail
to start ns_server back and not crash at the same time. That's because
it returns an error describing the situation to the caller. So we need
to explicitly handle this specific error.

Change-Id: I8caa7074999d25ab95f2e7335babcaed988586ce
Reviewed-on: http://review.couchbase.org/49296
Tested-by: Aliaksey Artamonau <aliaksiej.artamonau@gmail.com>
Reviewed-by: Aliaksey Artamonau <aliaksiej.artamonau@gmail.com>