MB-14841: don't setValue for rawStatsCell in case response 4xx
This bug occurs in case of onLoopData called with mark404 (which
is actually not only 404 but any 4xx). onLoopData is protected
from next call in case of async was canceled (this code doing
it !dataCallback.continuing(value)) but this issue is caused
by the same protective logic. In case async was not canceled
and _uistats returned with 4xx, dataCallback.continuing will
set empty object for rawStatsCell because mark404 is empty
object.

Change-Id: I365ce751806f7a1137b4755f347171b4d75426ee
Reviewed-on: http://review.couchbase.org/50758
Tested-by: Pavel Blagodov <stochmail@gmail.com>
Reviewed-by: Dave Finlay <dave.finlay@couchbase.com>
Tested-by: Dave Finlay <dave.finlay@couchbase.com>