Add a module for not per-bucket remote calls.
It's meant for global calls so we can't use janitor_agent which is
per-bucket. But the idea is similar, we want to have all our remote
calls to be listed in a single place.

I chose to use gen_server instead of just rpc-ing because by default
it spawn a process per request. I would like to have control over such
details. Currently the implementation is very simplistic (and there's
only one call to get indexes). But once (and if) we have more calls we
might think if something better is needed.

Change-Id: I1e5be1787a8369d9d4c76881df5ec40a7ab9168d
Reviewed-on: http://review.couchbase.org/51254
Tested-by: Dave Finlay <dave.finlay@couchbase.com>
Reviewed-by: Aliaksey Artamonau <aliaksiej.artamonau@gmail.com>
Tested-by: Aliaksey Artamonau <aliaksiej.artamonau@gmail.com>