MB-15107 include services into safe keys for merging node keys
...in mixed clusters

When you add already initialized node into cluster with some services
that are different that the services set up on that node the
{node, n1, services} get's changed on the node that adds n1

At the moment when modified {node, n1, services} key is merged on node n1,
the cluster_compat_mode for n1 is undefined. So n1 thinks that it's
pre 3.0 cluster and doesn't accept any config updates from the cluster.
Therefore the new value for services gets rejected.

Change-Id: Iffb0a4648f24b958d9ddb1a2a27d3b097803c1c0
Reviewed-on: http://review.couchbase.org/51733
Reviewed-by: Aliaksey Artamonau <aliaksiej.artamonau@gmail.com>
Tested-by: Aliaksey Artamonau <aliaksiej.artamonau@gmail.com>