MB-15354 Initialize _admin_pass again.
Recently I removed the support for upgrade from pre 2.5
versions. Unfortunately this had a side-effect of _admin_pass now being
always empty string. This is because it was actually set during the
chain of upgrades.

With this change I made it set as part of default config. Note that my
comment in the code stated that I didn't want it to be this way,
because ns_config_default:default/0 is supposed to be a pure
function. But we already violate this. And given that we hope to
replace ns_config soon, I opted with a simplest fix possible.

Change-Id: I6c8a246c983570c6f77da03127db6b7186348657
Reviewed-on: http://review.couchbase.org/52174
Tested-by: Aliaksey Artamonau <aliaksiej.artamonau@gmail.com>
Reviewed-by: Artem Stemkovski <artem@couchbase.com>