Grab master events in json format to master_events.log.
This also removes them from diag output. The motivation is that master
events in the format they were dumped to diag.log needed to be converted
to json first, before they could be fed to our scripts. So it saves a
lot of time to have them in json directly.

Change-Id: Ib1c6b5c8f94a418bd516a1c49b328c809a2e785d
Reviewed-on: http://review.couchbase.org/52675
Tested-by: Aliaksey Artamonau <aliaksiej.artamonau@gmail.com>
Reviewed-by: Artem Stemkovski <artem@couchbase.com>