Revert "MB-15926 Properly close last replication stream(s)."
This caused a regression in failover tests MB-16064. The reason for that
is that with the reverted change we'd wait for all streams to get closed
both on consumer and producer nodes. Since after failover we shut down
replications from the failed over node, closing producer streams on the
failed over node might take indefinitely long.

This reverts commit fe298eedea283502cca54e8d4990d93256de0260.

Change-Id: I14a28c7589f11a8262a4a56219da25900821b858
Reviewed-on: http://review.couchbase.org/54514
Reviewed-by: Poonam D <poonam@couchbase.com>
Tested-by: Aliaksey Artamonau <aliaksiej.artamonau@gmail.com>