Get rid of reformat task option.
It's purpose is to deal with windows tools returning output in utf-16
encoding. For that it used a subtle detail about python's print
statement, namely that it tries to convert a passed string to the
locale set for the file.

I continue using similar semantics of python's write on text files,
but I removed code duplication. So we have one code path handling all
the cases.

Change-Id: I900a8362304b8f504d461e46c04f085431181efc
Reviewed-on: http://review.couchbase.org/55511
Tested-by: Aliaksey Artamonau <aliaksiej.artamonau@gmail.com>
Reviewed-by: Artem Stemkovski <artem@couchbase.com>