MB-16435: OS X: Remove any overrides of DYLD_LIBRARY_PATH
As of OS X 10.11 (El Capitan) it is no longer possible to set /
override DYLD_LIBRARY_PATH. As such we need to move to embedding
runtime linker search paths in all required binaries.

Given we can't rely on this env var anymore, remove setting it in the
startup scripts to (1) ensure we don't incorrectly rely on it and (2)
help in validating that all uses are identified and fixed.

Change-Id: Ifb1b4da87239328e8f85f7a4011b3f2eecf8c44a
Reviewed-on: http://review.couchbase.org/56450
Reviewed-by: Dave Finlay <dave.finlay@couchbase.com>
Tested-by: Dave Finlay <dave.finlay@couchbase.com>