MB-16528 Activate delta recovery nodes before updating buckets.
Similarly, drop recovery_type and failover_vbuckets before updating
buckets. This is to enforce the invariant that node may be referenced
by the bucket configs only once it's active.

Without this change, if wait_for_bucket fails, node will continue to
be inactiveAdded, but bucket configs will have this node in some of
the chains. This breaks the re_failover logic that assumes that cannot
happen. So it's possible to re_failover such a node. It will get
ejected in the beginning of the next rebalance. But rebalance will try
to access this node to stop the replications.

Note that with the change even if, say, wait_for_bucket fails, even
though delta recovery nodes will have been activated by this point,
following rebalance would pickup deltaRecoveryMap from bucket config
and wait for the buckets activations (though with a shorter timeout).

Change-Id: Ib4804c2a086e3e276c492425cf333229895ceaf2
Reviewed-on: http://review.couchbase.org/56096
Tested-by: Aliaksey Artamonau <aliaksiej.artamonau@gmail.com>
Reviewed-by: Artem Stemkovski <artem@couchbase.com>