MB-16602: Support binary documents in the UI with a prefix of `_local/`
When there's a binary (non-JSON) document which has an ID that starts
with `_local/` it breaks the UI.

The reason is the call to `couch_doc:from_binary/3` which has a special
case for documents with a `_local/` prefix. The whole function isn't needed,
hence a simpler version of it called `capi_crud:is_valid_json/1` gets
introduced, that does only the JSON validation.

Change-Id: I1f1b5e3dbba543f792633f308c57a5782d872351
Reviewed-on: http://review.couchbase.org/56255
Reviewed-by: Artem Stemkovski <artem@couchbase.com>
Tested-by: Artem Stemkovski <artem@couchbase.com>