MB-16487 add ns_test as part of make test
... and remove ns_ui_test from being part of make test
The ui tests require updating an entire set of libraries starting
from c libraries (requirement to install chrome), which is probably
disruptive to the build process at this point. We need to figure
out a way to run these as part of the build (TBD currently)

Also change 'angular' to 'ui'.

Change-Id: I4cbf2f892758cdad9500f4f07ec93946b54205e7
Reviewed-on: http://review.couchbase.org/56799
Reviewed-by: Dave Finlay <dave.finlay@couchbase.com>
Tested-by: Dave Finlay <dave.finlay@couchbase.com>