MB-16677 Introduce service maps.
The semantic of a service map is as follows. It's a list of nodes that
have corresponding service running and are ready to handle client
requests. It must always be the case that get_service_map(Service) is a
subset of service_nodes(ns_node_disco:nodes_wanted(), Service). If it's
a proper subset, then rebalance is required to activate the service on
all configured nodes.

The purpose is to handle rebalance out more gracefully. Any node
service will first be deactivated in the corresponding service
map. After that ns_server (with some following changes) will stop
advertising the service on the node but will continue running the
actual process on the node. Eventually clients will detect new
configuration and will stop sending requests to the node. Then the
node will be removed in our normal fashion.

In more distant future service maps will also be used to track service
topologies for topology aware services (like cbft).

Change-Id: I20d79d64fe148687a0724816604a884f6b3de5f7
Reviewed-on: http://review.couchbase.org/56938
Tested-by: Aliaksey Artamonau <aliaksiej.artamonau@gmail.com>
Reviewed-by: Artem Stemkovski <artem@couchbase.com>