MB-16758: Add pluggable UI support to cluster_run
Adds the cluster_run option --pluggable-ui=File to make it possible to
spscify pluggable UI specification files that are in non-standard
locations. This is expected to be only used during development of new
services.

The files specified with this option has higher priority than files found
in the standard location. This makes it possible to override the normal
specifications for testing purposes.

Change-Id: Ica7c304c6fc7b2e643e9ea343b965570b7956b2f
Reviewed-on: http://review.couchbase.org/57179
Reviewed-by: Dave Finlay <dave.finlay@couchbase.com>
Tested-by: Anders Nygren <anders.nygren@couchbase.com>