MB-12222: avoid flickering of UI
In order to avoid flickering of UI we have to reload mnPoller
instead of .reloadState() and show spinner. This change intoduce
.broadcast method for mnPromiseHelper. Using $rootScope.$broadcast will
fire an event down the scope. $rootScope is the parent of all scopes,
which makes every newly created $scope a descendant, so for example we
can fire any event from modal windows. In other words
$rootScope.$broadcast will notify all $rootScope.$on as well as
$scope.$on.

Also we have to separate couple of states on sub states in order
to avoid flickering of UI while navigating in browser
using back/forward buttons:
  - app.admin.documents.control
  - app.admin.servers.list

And also we don't need anymore wrap state.go into _.defer since
it's related to .keepIn functionality.

Change-Id: Iead3b895c2cfa816f6733d9c611f230d77f73f85
Reviewed-on: http://review.couchbase.org/57843
Reviewed-by: Dave Finlay <dave.finlay@couchbase.com>
Tested-by: Dave Finlay <dave.finlay@couchbase.com>