MB-17648 return 403 and list of needed permissions if request
..is authenticated by not authorized

RFC 7231 (Hypertext Transfer Protocol (HTTP/1.1): Semantics and Content)

6.5.3. 403 Forbidden

The 403 (Forbidden) status code indicates that the server understood
the request but refuses to authorize it. A server that wishes to make
public why the request has been forbidden can describe that reason in
the response payload (if any).

If authentication credentials were provided in the request, the
server considers them insufficient to grant access. The client
SHOULD NOT automatically repeat the request with the same
credentials. The client MAY repeat the request with new or
different credentials. However, a request might be forbidden for reasons
unrelated to the credentials.

Change-Id: I9a0736bac4df742e690c90b65de778f1651ad585
Reviewed-on: http://review.couchbase.org/59150
Tested-by: Artem Stemkovski <artem@couchbase.com>
Reviewed-by: Aliaksey Artamonau <aliaksiej.artamonau@gmail.com>