MB-18093: make forestdb the default storage type
This change makes forestdb the default storage type during cluster
initialization and when index nodes are being added to the cluster
for the first time. This is a UI only change -- the only effect is
that the user doesn't have to select the appropriate radio button
if she wishes to pick up the default.

I left the cluster settings page alone for now and we can discuss
withPM and Rob on what to do there.

Change-Id: I9da353c77644a1fe18fc7292355fb21d77047f03
Reviewed-on: http://review.couchbase.org/59910
Tested-by: Dave Finlay <dave.finlay@couchbase.com>
Reviewed-by: Dave Finlay <dave.finlay@couchbase.com>