MB-17341: Get current replications from the replicators ...
... they have the most uptodate view.
Before we change the state of vbucket, we want to remove undesired
replications.
After the vbucket state change, new replications will be added
in apply_new_config_replicas_phase().
Earlier, remove_undesired_replications() was also adding new replications
in some cases.
The computation in remove_undesired_replications() is correct but it was
working on stale view of current replications. E.g. a stream might no longer
exist but replication_manager thinks it is. If that stream happened to be also
in FutureReps, then it was being added to the list passed
to set_desired_replications.

Change-Id: I73b12bf3037d8398d2cb54a0eb84eea410d04ed4
Reviewed-on: http://review.couchbase.org/60397
Tested-by: Poonam D <poonam@couchbase.com>
Reviewed-by: Aliaksey Artamonau <aliaksiej.artamonau@gmail.com>