MB-16872: change GOTRACEBACK to be "crash" by default
Existing users who are upgraded from 4.0 or 4.1 will get
the default if they have not previously modified the
GOTRACEBACK value. In particular if the value was modified
(say to equal crash) but then reset to empty string, these
users will not pick up the new default. But this is fine since
these users have explicitly set the value to "".

Change-Id: I01b87c67d841933bb408537cc52099ab7686cf5d
Reviewed-on: http://review.couchbase.org/61061
Tested-by: Dave Finlay <dave.finlay@couchbase.com>
Reviewed-by: Aliaksey Artamonau <aliaksiej.artamonau@gmail.com>