reduce amount of requests: broadcast mnPoolDefault events
Taking into consideration the fact that the mnPoolDefault
and mnTasksDetails are always relevant we can use cached
value of those services throughout the application. All we need
to know when the change is happened and then recalculate new value,
that is why we use broadcast in mnAdminController.

It reminds how Cell based application works but instead of reactive
values we use angular events. I would like to mention that we could use
scope.$watch instead of scope.broadcast/scope.on like
scope.$watch("poolDefault.serverGroupsUri", function () {
   //do something when serverGroupsUri changed
});
but I don't think that we should because:
1) it dependents on data in scope
2) we already use promiseHelper.broadcast in some places

benefits:
1) reduced amount of requests to the server
2) we know exactly when to fetch new data form the server instead of dumb periodic polling
3) now all tabs are sync, because mnPoolDefault is sync.

Change-Id: I2a79ed7799cc34885e12d5e775a4adfd86542950
Reviewed-on: http://review.couchbase.org/61499
Reviewed-by: Pavel Blagodov <stochmail@gmail.com>
Tested-by: Pavel Blagodov <stochmail@gmail.com>
Reviewed-by: Dave Finlay <dave.finlay@couchbase.com>
Tested-by: Dave Finlay <dave.finlay@couchbase.com>