MB-18289: actually minify app.min.js
This change actually minifies app.min.js. Up until this point
app.min.js was just a concatenation of the application javascript
files.

Minification is accomplished using Taco De Wolff's minification
library (https://github.com/tdewolff/minify), which is nice and
easy to use, is available under and MIT license and supports a
minification of several different kinds of web formats, obviously
including javascript.

Taco's minifier is similar in functionality to JSMin. See
https://github.com/tdewolff/minify#js-- for more details. No fancy
variable renaming is happening so savings are reduced, however,
so is risk. In our case prior to minifcation app.min.js came in at
about 544 kB. Post minification, it's about 397 kB; a 27% savings.
Not huge, but since we've come this far and it's so straightforward
to add, why not?

Change-Id: Iffc9d6038f35d9f482f21f96f717a94493eee138
Reviewed-on: http://review.couchbase.org/62474
Tested-by: Dave Finlay <dave.finlay@couchbase.com>
Reviewed-by: Aliaksey Artamonau <aliaksiej.artamonau@gmail.com>