MB-18775, MB-17571: default index and fts mem quota to 512 MB
Fixing both of these issues with the following strategy:
1) KV mem quota is unchanged (60% of max avail memory on the machine
   of 256 MB whichever is higher)
2) Index quota is suggested at 512 MB by default (min of 256 MB)
3) FTS quota default limits at 512 MB but it otherwise calculated
   as it is today (20% of the available after the KV quota is
   subtracted maxing out at 512.) For reference, in deployments
   where the max available memory is > 7 GB the 512 limit for
   default FTS quota will be hit.

Of course these are just defaults and the user is free to adjust
them as he or she sees fit. The problem remains that in deployments
with very little available memory (< 2 GB) the suggested defaults
will sum to greater than available memory. But because this is
an unusual situation these days, I figure it's easiest to simply
let this be.

Change-Id: I96f04763f9b7bd942e5fce8bb71acb30b7ec4444
Reviewed-on: http://review.couchbase.org/61858
Tested-by: Dave Finlay <dave.finlay@couchbase.com>
Reviewed-by: Aliaksey Artamonau <aliaksiej.artamonau@gmail.com>