Try working around the problem where global gets out of sync.
In all the logs that I examined, the global_name_server was in the
state where there some nodes that are known but are not synced. And it
doesn't seem to attempt to resync with such nodes. I still don't know
what the actual problem is. Instead this patch tries to detect such
condition and forces reconnect to the unsynced nodes in the hope that
it will resolve the problem.

This behavior seems safe. But it can be disabled anyway by setting
enable_global_workaround to false in ns_config.

Change-Id: I55d182dbccbdc5a73b7c7c2256fd1d2e505d561f
Reviewed-on: http://review.couchbase.org/62010
Tested-by: Aliaksey Artamonau <aliaksiej.artamonau@gmail.com>
Reviewed-by: Artem Stemkovski <artem@couchbase.com>