MB-18307: increase editable doc size limit to 256 KiB
Given that we've saved a bunch of memory with the angular
minification, I thought we should spend it on documents. :-)
(That's a joke, of course.)

Simply increased the limit and change the warning message.
Tested it on Chrome, Firefox, Safari on Mac and Windows 8 and
it seems to work fine. Firefox seemed to be a tiny bit sluggish
with the scroll, but was very functional, so I figure let's do
it.

Change-Id: I285dea61549a0f6d6239407d2ae51440f2c863a1
Reviewed-on: http://review.couchbase.org/62269
Tested-by: Dave Finlay <dave.finlay@couchbase.com>
Reviewed-by: Pavel Blagodov <stochmail@gmail.com>