Support "minified" query param against index.html.
This change is to help us debug situations where a user is having
a UI problem and we suspect it may be caused by minification.

The idea is as follows: if minified is specified as a query
parameter when vending index.html, then it's value is honored
(minified =:= "true" returns minifed, anything else is non-
minified.) If minified is not specified as a query paramater
it falls back to the application environment variable and
default.

This parameter won't be advertized; it's simply to help quickly
ascertain if a problem is being caused by minification. Parameter
only works against index.html (if you append it to a URL once
you've brought up the UI, it doesn't have any affect.)

Change-Id: Ic5623b4e04319ef1793a501dc9ba44e7e7b182c9
Reviewed-on: http://review.couchbase.org/62274
Tested-by: Dave Finlay <dave.finlay@couchbase.com>
Reviewed-by: Aliaksey Artamonau <aliaksiej.artamonau@gmail.com>