MB-18623 - change implementation of avg_queries_latency
Previously we computed the average in cbft and it represented
a server lifetime average.  This was inconsistent with the other
similar stats shown by other services.

Now we return query time and query count as counters, and let
ns_server compute the avg across the most recent sample.  This is
consistent with how the query and index service show similar
stats.

Change-Id: I140da0c79e949e98f2be1bd5559056b55f3e4aad
Reviewed-on: http://review.couchbase.org/62295
Tested-by: Marty Schoch <marty.schoch@gmail.com>
Reviewed-by: Steve Yen <steve.yen@gmail.com>
Reviewed-by: Artem Stemkovski <artem@couchbase.com>