MB-19105: "redirect" user to destination after login
This change does the following.
(1) transitions to the app.auth state now in certain
    cases instruct the URL not to change (using the
    {location:false} mechansim in ui-router
(2) post successful login, if the URL in the URL bar is
    not the auth URL, the auth controller will sync to
    the URL at which point the normal angular routing
    goodness will kick in, sending the user to the right
    part of the app.

Ultimately, this has the effect that we want: a user can
paste a complex analytics URL into the URL bar on his  or
her browser and post login will be taken to that exact
graph.

Change-Id: I78a1bf3662df12fc50192939aa1cba38c6aa6cdf
Reviewed-on: http://review.couchbase.org/62549
Reviewed-by: Dave Finlay <dave.finlay@couchbase.com>
Tested-by: Dave Finlay <dave.finlay@couchbase.com>