MB-19134: don't open views link in new tab
The behavior of the Views, Query and FTS links should be consistent when
it comes to redirecting a user to avail of the service on a different node.
I think it's best if we use the same tab rather than open a new tab -
I think it's less interrupting for the user. Anyway, as such the
target=_blank is removed.

Change-Id: Iac0367c35e5619fe2f4b30ea880275e3d980cf1e
Reviewed-on: http://review.couchbase.org/62752
Tested-by: Dave Finlay <dave.finlay@couchbase.com>
Reviewed-by: Aliaksey Artamonau <aliaksiej.artamonau@gmail.com>