MB-19037 merge_kv_pairs to return a set of keys to rereplicate.
This is needed for the keys whose vclock changes as a result of
conflict resolution. Previously, this wouldn't trigger replication. So
even though all nodes would have the same value for the conflicting
keys, the vclocks would still indicate a conflict. So if the same key
is changed later by any of the nodes, it would trigger an unnecessary
conflict.

This patch doesn't change the replication logic just yet, it only
extends ns_config:merge_kv_pairs.

Change-Id: Ifd006fa1dc09504e06e67fd817f76e76d2e81e49
Reviewed-on: http://review.couchbase.org/63956
Reviewed-by: Artem Stemkovski <artem@couchbase.com>
Tested-by: Aliaksey Artamonau <aliaksiej.artamonau@gmail.com>