Get rid of replace version of cas_config.
The only difference between 'replace' and 'remote' behavior is that the
former announces changes on ns_config_events_local. The purpose of this
is to replicate locally made changes to other nodes. Since
ns_couchdb (the only user of 'replace') doesn't replicate its config
anywhere, there's no need for this behavior.

Change-Id: I1af5897e3093d0e7ea65ffcf14df218908287996
Reviewed-on: http://review.couchbase.org/63958
Tested-by: Aliaksey Artamonau <aliaksiej.artamonau@gmail.com>
Reviewed-by: Artem Stemkovski <artem@couchbase.com>