MB-19677 Skip compaction for moved out vbuckets.
Since we get the list of vbuckets to compact in the very beginning of
compaction iteration, it's possible that by the time we get to
compacting some of them, they are already moved out from the
node. With this change we don't pollute UI log with unnecessary
"compaction failed" messages.

Change-Id: I59a9316cf69516a9e2f2b4ca44976c970822a86d
Reviewed-on: http://review.couchbase.org/64347
Tested-by: Aliaksey Artamonau <aliaksiej.artamonau@gmail.com>
Reviewed-by: Artem Stemkovski <artem@couchbase.com>